refactor docker helper away from lint
