+++ demisto_sdk/commands/common/MDXServer.py
@@ -10,9 +10,9 @@ import docker.errors
 import docker.models.containers
 
 from demisto_sdk.commands.common.constants import DEPENDENCIES_DOCKER
+from demisto_sdk.commands.common.docker_helper import (
+    Docker, init_global_docker_client)
 from demisto_sdk.commands.common.errors import Errors
-from demisto_sdk.commands.lint.docker_helper import (Docker,
-                                                     init_global_docker_client)
 
 EXPECTED_SUCCESS_MESSAGE = 'MDX server is listening on port'
