+++ demisto_sdk/commands/common/docker_helper.py
