+++ demisto_sdk/commands/common/hook_validations/readme.py
@@ -18,6 +18,7 @@ from urllib3.util import Retry
 
 from demisto_sdk.commands.common.constants import (RELATIVE_HREF_URL_REGEX,
                                                    RELATIVE_MARKDOWN_URL_REGEX)
+from demisto_sdk.commands.common.docker_helper import init_global_docker_client
 from demisto_sdk.commands.common.errors import (FOUND_FILES_AND_ERRORS,
                                                 FOUND_FILES_AND_IGNORED_ERRORS,
                                                 Errors)
@@ -31,7 +32,6 @@ from demisto_sdk.commands.common.tools import (
     compare_context_path_in_yml_and_readme, get_content_path,
     get_url_with_retries, get_yaml, get_yml_paths_in_dir, print_warning,
     run_command_os)
-from demisto_sdk.commands.lint.docker_helper import init_global_docker_client
 
 json = JSON_Handler()
