+++ demisto_sdk/commands/lint/helpers.py
@@ -25,7 +25,7 @@ from packaging.version import parse
 # Local packages
 from demisto_sdk.commands.common.constants import (TYPE_PWSH, TYPE_PYTHON,
                                                    DemistoException)
-from demisto_sdk.commands.lint.docker_helper import init_global_docker_client
+from demisto_sdk.commands.common.docker_helper import init_global_docker_client
 
 # Python2 requirements
 PYTHON2_REQ = ["flake8", "vulture"]
