+++ tests_end_to_end/test_dockerserver.py
@@ -2,9 +2,9 @@ import requests
 from requests.sessions import HTTPAdapter
 from urllib3 import Retry
 
+from demisto_sdk.commands.common.docker_helper import init_global_docker_client
 from demisto_sdk.commands.common.MDXServer import (DEMISTO_DEPS_DOCKER_NAME,
                                                    start_docker_MDX_server)
-from demisto_sdk.commands.lint.docker_helper import init_global_docker_client
 
 SERVER_DNS = 'docker'  # switch to localhost for testing locally
